# bajaj-admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medic 24x7 App",
        "description": "REST API's for Medic24x7 application",
        "termsOfService": "",
        "contact": {
            "email": "sayalib@leometric.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "CUSTOMER PROFILE"
                ],
                "summary": "Get User Profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/profile": {
            "get": {
                "tags": [
                    "PARTNER PROFILE"
                ],
                "summary": "Get Partner Profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}